# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hookah Club subscription management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TelegramAuthInput {
  telegramId: number;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  staff: 'staff',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  telegramId: number;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  role: UserRole;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  displayCode?: string | null;
  createdAt: string;
  loyaltyStamps: number;
  loyaltyTotalRedeemed: number;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UpdateNoteInput {
  /** @nullable */
  note: string | null;
}

export interface SubscriptionPlan {
  id: number;
  nameRu: string;
  nameRs: string;
  level: number;
  hookahCount: number;
  priceRsd: number;
  pricePerHookah: number;
  bonusHookahFruit: number;
  bonusElectric: number;
  bonusHookahCheap: number;
}

export interface UserSubscriptionDetail {
  id: number;
  userId: number;
  planId: number;
  plan: SubscriptionPlan;
  hookahsRemaining: number;
  fruitHookahsRemaining: number;
  electricAvailable: boolean;
  cheapHookahAvailable: boolean;
  activatedAt: string;
  /** @nullable */
  expiresAt: string | null;
  /** @nullable */
  note?: string | null;
  isLegacy: boolean;
}

export interface SubscriptionActivateInput {
  planId: number;
  /** @nullable */
  note?: string | null;
  isLegacy?: boolean;
}

export interface SubscriptionUsageUpdate {
  /** @nullable */
  hookahsRemaining?: number | null;
  /** @nullable */
  fruitHookahsRemaining?: number | null;
  /** @nullable */
  electricAvailable?: boolean | null;
  /** @nullable */
  cheapHookahAvailable?: boolean | null;
  /** @nullable */
  note?: string | null;
}

export type UpdateUserRoleInputRole = typeof UpdateUserRoleInputRole[keyof typeof UpdateUserRoleInputRole];


export const UpdateUserRoleInputRole = {
  user: 'user',
  staff: 'staff',
  admin: 'admin',
} as const;

export interface UpdateUserRoleInput {
  role: UpdateUserRoleInputRole;
}

export interface AdminUserView {
  id: number;
  telegramId: number;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  role: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  displayCode?: string | null;
  createdAt: string;
  loyaltyStamps: number;
  loyaltyTotalRedeemed: number;
  subscription?: UserSubscriptionDetail;
}

export interface ActionLog {
  id: number;
  staffId: number;
  guestId: number;
  /** @nullable */
  staffName?: string | null;
  /** @nullable */
  guestName?: string | null;
  action: string;
  description: string;
  createdAt: string;
}

export interface LoyaltyStatus {
  loyaltyStamps: number;
  loyaltyTotalRedeemed: number;
}

export type AdminStatsSubscriptionsByPlanItem = {
  planName: string;
  activeCount: number;
  totalEver: number;
};

export interface AdminStats {
  totalUsers: number;
  activeSubscriptions: number;
  totalHookahsUsed: number;
  totalFruitUsed: number;
  totalCheapUsed: number;
  totalElectricUsed: number;
  totalActivations: number;
  legacyActiveCount: number;
  subscriptionsByPlan?: AdminStatsSubscriptionsByPlanItem[];
}