# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hookah Club subscription management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/telegram:
    post:
      operationId: authTelegram
      tags: [auth]
      summary: Authenticate via Telegram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramAuthInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /users/me/note:
    patch:
      operationId: updateMyNote
      tags: [users]
      summary: Update my personal note (guest)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteInput"
      responses:
        "200":
          description: Note updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    TelegramAuthInput:
      type: object
      required: [telegramId, firstName]
      properties:
        telegramId:
          type: integer
        firstName:
          type: string
        lastName:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    UpdateNoteInput:
      type: object
      required: [note]
      properties:
        note:
          type: ["string", "null"]
    User:
      type: object
      required: [id, telegramId, firstName, role, createdAt, loyaltyStamps, loyaltyTotalRedeemed]
      properties:
        id:
          type: integer
        telegramId:
          type: integer
        firstName:
          type: string
        lastName:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, staff, admin]
        note:
          type: ["string", "null"]
        displayCode:
          type: ["string", "null"]
        createdAt:
          type: string
        loyaltyStamps:
          type: integer
        loyaltyTotalRedeemed:
          type: integer
...