# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hookah Club subscription management API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  telegramId: number;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  role: UserRole;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  displayCode?: string | null;
  createdAt: string;
  loyaltyStamps: number;
  loyaltyTotalRedeemed: number;
}